+++ hs/abdata_container.hs
@@ -1531,7 +1531,7 @@ var val : ”‚¦‚é’l
 
 ;--------------------
 %index
-container_isValid
+container_is_valid
 container —v‘f”Ô†‚ª—LŒø‚©‚Ç‚¤‚©
 
 %prm